SALSA: Fix Byte Grid handling
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@9897 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: 39b67cca5b9b3f9fbb87277498f1b2899340f71f

+++ application/SaLsA/de.offis.salsa.playground/src/de/uniol/inf/is/odysseus/salsa/ui/GridScreen.java
@@ -37,7 +37,7 @@ public class GridScreen extends JFrame {
         this.setDefaultCloseOperation(HIDE_ON_CLOSE);
     }
 
-    public void onGrid(final Double[][] grid) {
+    public void onGrid(final Byte[][] grid) {
         this.map.onGrid(grid);
     }
 }
